scaler/requirements.txt
grpcio==1.62.1
protobuf==4.25.3
vastai==0.2.0
docker==7.0.0